Refactor directory structure for agents and services
