+++ jarvis/__init__.py
@@ -1,6 +1,6 @@
 """Jarvis calendar assistant package."""
 
-from .calendar_service import CalendarService
+from .services.calendar_service import CalendarService
 from .logger import JarvisLogger
 from .log_viewer import LogViewer
 from .ai_clients import (
@@ -9,9 +9,9 @@ from .ai_clients import (
     OpenAIClient,
     AnthropicClient,
 )
-from .network import AgentNetwork
-from .network.agents.calendary_agent import CollaborativeCalendarAgent
-from .network.agents.ui_agent import UIAgent
+from .agents.agent_network import AgentNetwork
+from .agents.calendar_agent import CollaborativeCalendarAgent
+from .agents.ui_agent import UIAgent
 from .main_network import JarvisSystem, create_collaborative_jarvis
 
 __all__ = [
