--- jarvis/network/agents/__init__.py
@@ -1,6 +1,0 @@
-"""Agent implementations used within the AgentNetwork."""
-
-from .calendary_agent import CollaborativeCalendarAgent
-from .ui_agent import UIAgent
-
-__all__ = ["CollaborativeCalendarAgent", "UIAgent"]
