+++ jarvis/network/core.py
@@ -1,8 +1,8 @@
 """Legacy imports for backward compatibility."""
-from .message import Message
-from .capability import Capability
-from .agent_network import AgentNetwork
-from .base_agent import NetworkAgent
+from ..agents.message import Message
+from ..agents.capability import Capability
+from ..agents.agent_network import AgentNetwork
+from ..agents.base import NetworkAgent
 
 __all__ = [
     "Message",
